Bluetooth: Audio: Media control - Test function - invalid search length
Add a shell command to reproduce a search control point hard fault.

This shell command was created to reproduce a hard fault observed
during IOP testing.  The issue was that the data contained a zero
length SCI field, from which 1 was subtracted and the result used in a
memcpy.

The issue has been fixed, the command is kept to test others.

Only compiled in if CONFIG_BT_DEBUG_MCS and CONFIG_BT_TESTING are
defined.

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>