Bluetooth: Audio: Media control - change to Generic MCS
Change to use the Generic Media Control Service instead of the Media
Control Service.

The spec says that "The Media Control Server shall instantiate zero or
more MCS and a single GMCS."

Previously, this implementation has had a single MCS, and no
GMCS. After this change, this implementation has a single GMCS, and is
compliant with the spec in that respect.

- Use the GMCS UUID for the service
- Discover GMCS instead of MCS in client
- Make the player name notifiable in service (mandatory for GMCS)
- Handle player name notifications in client

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>