Bluetooth: Audio: Build without Object Transfer Service client
Make it possible to build the audio shell when the Object Transfer
Service client (OTC) is not enabled.

The affected parts are the Media Control Service client (MCC) and the
shell commands for MCC.  The change is done by adding "#ifdef-guards"
around the OTS-related parts of the code and slight restructuring.

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>