Bluetooth: Audio: Changed AICS index to pointer
Changed how the individual AICS instances are handled.
Instead of using an index, it is now a pointer to
the specific instance. A server/client may now
register an instance, and use the pointer in future
AICS interactions, rather than an index.

This fixes some issues found with VCS/MICS and indexes,
while also simplifying how AICS is used by multiple
other services.

This does not update MICS/VCS to be pointer based in
the API yet, as that will be done in a separate commit.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>