Bluetooth: Audio: Changed VOCS index to pointer
Changed how the individual VOCS instances are handled.
Instead of using an index, it is now a pointer to
the specific instance. A server/client may now
register an instance, and use the pointer in future
VOCS interactions, rather than an index.

The VCS API has also been updated to reflect these,
and has a function that returns the necessary pointers
added.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>