Bluetooth: Audio: VCS and MICS only use public API for AICS
VCS and MICS used to use an internal header for AICS. Since
it would be nice if other application services could use AICS
in the same way, the public API need to provide what is needed
to support that. The public header for AICS now does that and
VCS and MICS have been updated to use that exclusively.

Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>