Bluetooth: Audio: Media control - reorder test code - no changes
Order test code (switch the order of two parts) so that the testing
follows the ordering of characteristics in the service spec.  (This
ordering is also followed in the client and server implementations.)

(The diff looks like a lot of changes, but that is because the two
parts switched are very similar, so the diff does not pick up that
this is just a move of one part.  Except for the moving of code, there
are no changes here.)

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>