Bluetooth: Audio: Media control - remove DEBUG-dependency for callbacks
Remove the dependency on DEBUG for object selected and object metadata
read callbacks.

The original thought was that these callbacks were not required for
the shell/application.  But they are now being used in the tests, so
the tests will fail if debugging is turned off.

This behaviour may be changed in the upcoming API change - these
callbacks may be removed from the public API then.

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>