Bluetooth: Audio: Media control - modify test flag handling
Change to clearing flag just before they will be used, rather than
clearing them afterwards and relying on the previous user have cleared
them and the flags not having been set again in the meantime.

This helps solve the issue where the same flag is set multiple time
due to the same callback being used both for reads and for
notifications.  For some characteristics, notifications are arriving
as a result of other tests, in which case the flag may be set already
before the next read of that characteristic.  In which case the test
will proceed before the intended operation has completed.  Clearing
the flag just before the operation prevents that.

Signed-off-by: Asbjørn Sæbø <asbjorn.sabo@nordicsemi.no>